requirements.txt
huggingface-hub==0.19.4
sagemaker==2.199.0
pypdfium2==4.24.0
httplib2==0.8
langchain==0.0.340
requests==2.31.0
pandas==2.1.3
boto3==1.33.3
pillow==10.1.0
faiss-cpu==1.7.4
opensearch-py
numexpr==2.8.0